Remove unused Redux actions and reducers and rename homepage component
