+++ assets/js/index.js
@@ -1,2 +1,1 @@
-// import pages
-import 'pages/homePage';
+import 'pages/HomePage';
