--- assets/js/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import auth from './auth';
-
-// WIP
-
-const pokebattleReducer = combineReducers({
-  auth,
-});
-
-export default pokebattleReducer;
